# [BREAKING] Change Processor interface to img instead of []byte to improve performance