--- a/pkg/processor/native/processor_test.go
+++ b/pkg/processor/native/processor_test.go
@@ -1,7 +1,6 @@
 package native
 
 import (
-	"bytes"
 	"github.com/gojek/darkroom/pkg/processor"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
@@ -13,16 +12,20 @@ import (
 type BildProcessorSuite struct {
 	suite.Suite
 	srcData       []byte
+	srcImage      image.Image
 	watermarkData []byte
 	badData       []byte
+	badImage      image.Image
 	processor     processor.Processor
 }
 
 func (s *BildProcessorSuite) SetupSuite() {
 	s.processor = NewBildProcessor()
 	s.srcData, _ = ioutil.ReadFile("_testdata/test.png")
+	s.srcImage, _, _ = s.processor.Decode(s.srcData)
 	s.watermarkData, _ = ioutil.ReadFile("_testdata/overlay.png")
 	s.badData = []byte("badImage.ext")
+
 }
 
 func TestNewBildProcessor(t *testing.T) {
@@ -30,31 +33,27 @@ func TestNewBildProcessor(t *testing.T) {
 }
 
 func (s *BildProcessorSuite) TestBildProcessor_Resize() {
-	output, err := s.processor.Resize(s.srcData, 500, 500)
+	out := s.processor.Resize(s.srcImage, 500, 500)
 
-	assert.NotNil(s.T(), output)
-	assert.Nil(s.T(), err)
-
-	img, _, _ := image.Decode(bytes.NewReader(output))
-	assert.Equal(s.T(), 500, img.Bounds().Dx())
-	assert.Equal(s.T(), 375, img.Bounds().Dy())
+	assert.NotNil(s.T(), out)
+	assert.Equal(s.T(), 500, out.Bounds().Dx())
+	assert.Equal(s.T(), 375, out.Bounds().Dy())
 }
 
 func (s *BildProcessorSuite) TestBildProcessor_Crop() {
-	output, err := s.processor.Crop(s.srcData, 500, 500, processor.CropCenter)
+	out := s.processor.Crop(s.srcImage, 500, 500, processor.CropCenter)
 
-	assert.NotNil(s.T(), output)
-	assert.Nil(s.T(), err)
+	assert.NotNil(s.T(), out)
 
-	img, _, _ := image.Decode(bytes.NewReader(output))
-	assert.Equal(s.T(), 500, img.Bounds().Dx())
-	assert.Equal(s.T(), 500, img.Bounds().Dy())
+	assert.Equal(s.T(), 500, out.Bounds().Dx())
+	assert.Equal(s.T(), 500, out.Bounds().Dy())
 }
 
 func (s *BildProcessorSuite) TestBildProcessor_Grayscale() {
 	var actual, expected []byte
 	var err error
-	actual, err = s.processor.GrayScale(s.srcData)
+	out := s.processor.GrayScale(s.srcImage)
+	actual, err = s.processor.Encode(out, "png")
 	assert.NotNil(s.T(), actual)
 	assert.Nil(s.T(), err)
 
@@ -75,19 +74,7 @@ func (s *BildProcessorSuite) TestBildProcessor_Watermark() {
 }
 
 func (s *BildProcessorSuite) TestBildProcessorWithBadInput() {
-	output, err := s.processor.Crop(s.badData, 0, 0, processor.CropCenter)
-	assert.NotNil(s.T(), err)
-	assert.Nil(s.T(), output)
-
-	output, err = s.processor.Resize(s.badData, 0, 0)
-	assert.NotNil(s.T(), err)
-	assert.Nil(s.T(), output)
-
-	output, err = s.processor.GrayScale(s.badData)
-	assert.NotNil(s.T(), err)
-	assert.Nil(s.T(), output)
-
-	output, err = s.processor.Watermark(s.badData, s.watermarkData, 255)
+	output, err := s.processor.Watermark(s.badData, s.watermarkData, 255)
 	assert.NotNil(s.T(), err)
 	assert.Nil(s.T(), output)