**Changes:**
This PR is a breaking change to the Processor interface.
- Change Processor interface to `image.Image` instead of `[]byte`
- Remove unneccessary cloning in `grayScale` method

**Reason for changes**:
- 15% performance increase on manipulator.Process when given params is more than one operation (resize/crop + grayscaling).

**Profiling details:**
The profiling is done by processing 96 images from Image Manipulation dataset (https://www5.cs.fau.de/research/data/image-manipulation/) which have varying file sizes (they also have >10 MB pictures).
The process being done is cropping to 50% of width and height with right anchor and grayscaling

**Profiling result:**
using `[]byte` interface:

```
Total: 58.286914s
Average: 0.607s
Minimum: 0.024093s
Maximum: 1.910421s
```

using `image.Image` interface:
```
Total time: 49.394539s
Average: 0.515s
Minimum: 0.018555s
Maximum: 0.998278s
```