@adalberht 2019-07-04
Working on it.
I'll give the memory and cpu profiling result here later